Converted PlanetData code to double precision

+++ src/planet/face.cpp
@@ -106,10 +106,10 @@ std::array<unsigned int, 3> Face::getVertexIndices() const {
 	return this->vertexIndices;
 }
 
-void Face::calculateMidpoint(const std::vector<glm::vec3>& vertices) {
+void Face::calculateMidpoint(const std::vector<glm::dvec3>& vertices) {
 		/// We calculate the midpoint by averaging all vertex positions
 		this->midpoint = (vertices[this->vertexIndices[0]] +
 						  vertices[this->vertexIndices[0]] +
-						  vertices[this->vertexIndices[0]]) / 3.0f;
+						  vertices[this->vertexIndices[0]]) / 3.0;
 }
 } /// namespace lillugsi::planet
