+++ src/planet/vertexdata.h
@@ -12,7 +12,7 @@ namespace lillugsi::planet {
 class VertexData {
 public:
 	/// Create a vertex at the given position with default elevation of 0
-	explicit VertexData(const glm::vec3& position);
+	explicit VertexData(const glm::dvec3& position);
 
 	/// Rule of five - we manage complex relationships between vertices
 	~VertexData() = default;
@@ -22,14 +22,14 @@ public:
 	VertexData& operator=(VertexData&& other) noexcept = default;
 
 	/// Elevation accessors
-	[[nodiscard]] float getElevation() const { return elevation; }
-	void setElevation(float newElevation);
+	[[nodiscard]] double getElevation() const { return elevation; }
+	void setElevation(double newElevation);
 
 	/// Position accessors
-	[[nodiscard]] glm::vec3 getPosition() const { return position; }
+	[[nodiscard]] glm::dvec3 getPosition() const { return position; }
 	/// Get the current normal vector, recalculating if necessary
 	/// This provides an efficient way to access the normal while ensuring it's up to date
-	[[nodiscard]] glm::vec3 getNormal();
+	[[nodiscard]] glm::dvec3 getNormal();
 
 	/// Neighbor management
 	/// We store weak_ptrs to neighbors to avoid circular reference issues
@@ -38,7 +38,7 @@ public:
 
 	/// Get the slope to a specific neighbor
 	/// Calculates and caches the slope if needed
-	[[nodiscard]] float getSlope(size_t neighborIndex);
+	[[nodiscard]] double getSlope(size_t neighborIndex);
 
 	/// Force recalculation of normal vector based on neighbor positions
 	void recalculateNormal();
@@ -68,22 +68,22 @@ private:
 
 	/// Calculate initial distance to a neighbor
 	/// Called once during neighbor setup since lateral positions don't change
-	[[nodiscard]] float calculateDistanceToNeighbor(const VertexData& neighbor) const;
+	[[nodiscard]] double calculateDistanceToNeighbor(const VertexData& neighbor) const;
 
 	/// Physical properties
-	float elevation{-2.0f};
+	double elevation{-2.0};
 
 	/// Vector fields
-	glm::vec3 position;
-	glm::vec3 normal{0.0f, 1.0f, 0.0f};  /// Default normal points up
+	glm::dvec3 position;
+	glm::dvec3 normal{0.0, 1.0, 0.0};  /// Default normal points up
 
 	/// Topology and cached calculations
 	std::vector<std::weak_ptr<VertexData>> neighbors;
 	/// Store distances to neighbors - these remain constant after initialization
 	/// since only elevation changes, not lateral positions
-	std::vector<float> neighborDistances;
+	std::vector<double> neighborDistances;
 	/// Cache calculated slopes to avoid repeated calculations
-	std::vector<float> neighborSlopes;
+	std::vector<double> neighborSlopes;
 	/// Track which slopes need recalculation due to elevation changes
 	std::vector<bool> slopeDirtyFlags;
 	/// Track if the normal needs recalculation due to elevation changes
@@ -91,7 +91,7 @@ private:
 	bool normalDirty{true};
 
 	/// Constant used for floating point comparisons
-	static constexpr float EPSILON = 0.0000001f;
+	static constexpr double EPSILON = 0.0000001;
 };
 
 } /// namespace lillugsi::planet
